ai/requirements.txt
# Core API
fastapi
uvicorn
pydantic
python-multipart
python-dotenv

# AI & Image Processing
ultralytics         # YOLOv8 Object Detection
opencv-python       # Image Processing (cv2)
Pillow              # Image handling
numpy               # Math & Array handling
torch               # CLIP 내부 의존성 (transformers 사용 시 필수)

# Vector Search
pinecone>=3.0.0     # v3+ SDK (response.matches 방식 사용)
transformers

# LLM
google-generativeai
